Rely on egrep and regex to create a AUTHOR file blacklist
When generating the AUTHORS file the Makefile relies on a git log that may or may not be representative of the valuable data. It's very easy to omit configuration of the git config which easily makes for duplicate authors.

This change tweaks the output of the sorted git log, pipes the values into egrep that relies on an inverse filter that's easily extended. This change should not be seen as a long-term solution as the regex could quickly grow out of hand.