Pass in host user/group ID
When building kubicorn on CI systems, it's helpful for the resulting
binary to be owned by the CI user, so that it can be deleted during
clean up without permission issues. Pass in the current user/group
ID to Docker to facilitate this.

It is not passed in to the shell target, as the user is almost
guaranteed to not exist on the image, and a valid user should be present
when working in the shell.

Signed-off-by: Nolan Brubaker <nolan@heptio.com>