Enable AWS cloud provider integration on Ubuntu
In order to provider AWS cloud resources to Kubernetes, the kubelets
need to know about the cloud provider, as well as the apiserver and
controller-manager.

These updates will write the argument for the kubelet to use the cloud
provider, and pass the cloud provider information through kubeadm's
configuration on the master for the rest. The configration file
completely replaces the command arguments due to kubeadm not allowing
mixes of config files and arguments.

Additionally, node names must be the fully qualified host name on AWS in
order to properly register with the master. kubeadm is bumped to version
1.8.4 to support this.

Finally, in order to ease use, install the standard AWS storage class
into the cluster.

Signed-off-by: Nolan Brubaker <nolan@heptio.com>