Use an existing instance profile if found
Since we can detect whether or not an instance profile exists via error
types, attempt to re-use an existing instance profile before returning
an error.

The name and id have to be extracted seperately since the returned
values for creating an instance profile and getting one are two
different types.

Signed-off-by: Nolan Brubaker <nolan@heptio.com>